Full PullParser Support added
Added the Currency class in order to store the information from the RSS feed, as well as an ArrayList to store the items in.

Added a full section for PullParser which takes the relevant information - Title, Description, Publishing Date - and stores them in a Currency object.

Three attributes are stored as Strings currently but this will likely be changed for future functionality.