Modified Data Reading
Currency now has a double attribute: rate

This saves the value of 1GBP in local currency.

Modified pullparser code so that it uses substring methods to extract the exchange rate from the description string.

Modified startProgress() to execute on Creation for debug purposes.

Removed startButton()

Modified and more neatly formatted the activity_main file so that there is a slight margin at the top of the screen. Looks neater and increases compatibility with phones that have front cameras on the screen.